+++ src/main/resources/templates/admin-home-page.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>Library ADMIN</title>
+</head>
+<body>
+    <nav>
+        <ul>
+            <li><a th:href="@{/admin/users}">Users list</a></li>
+            <li><a th:href="@{/admin/users/test}">Specific user called test</a></li>
+            <li><a th:href="@{/admin/edit-profile/test}">Edit specific user called test</a></li>
+            <li><a th:href="@{/books/allbooks}">All Donated books</a></li>
+        </ul>
+    </nav>
+</body>
+</html>
